requirement.txt
pymupdf
sentence-transformers
faiss-cpu
torch
streamlit
transformers